use ts for mobile analytics, logging updates
