+++ src/desktop/assets/analytics.ts
@@ -1,7 +1,6 @@
 import $ from "jquery"
 import { data as sd } from "sharify"
-import { extend, pick } from "lodash"
-import { trackEvent } from "desktop/analytics/helpers"
+import { onAnalyticsReady, trackEvent } from "desktop/analytics/helpers"
 const analyticsHooks = require("../lib/analytics_hooks.coffee")
 const mediator = require("../lib/mediator.coffee")
 const setupSplitTests = require("../components/split_test/setup.coffee")
@@ -42,26 +41,7 @@ require("../analytics/main_layout.ts")
 
 $(() =>
   window.analytics.ready(function () {
-    if (sd.CURRENT_USER != null ? sd.CURRENT_USER.id : undefined) {
-      const allowedlist = [
-        "collector_level",
-        "default_profile_id",
-        "email",
-        "id",
-        "name",
-        "phone",
-        "type",
-      ]
-      const traits = extend(pick(sd.CURRENT_USER, allowedlist), {
-        session_id: sd.SESSION_ID,
-      })
-      window.analytics.identify(sd.CURRENT_USER.id, traits, {
-        integrations: { Marketo: false },
-      })
-      // clear analytics cache when user logs out
-      analyticsHooks.on("auth:logged-out", () => window.analytics.reset())
-    }
-
+    onAnalyticsReady()
     setupSplitTests()
 
     require("../analytics/global.js")
